+++ crates/filter/src/lib.rs
@@ -6,8 +6,9 @@
 //! 2. sort the all lines with a match result.
 //! 3. print the top rated filtered lines to stdout.
 
-mod source;
-mod worker;
+mod parallel_worker;
+mod sequential_source;
+mod sequential_worker;
 
 use std::sync::Arc;
 
@@ -16,14 +17,28 @@ use rayon::prelude::*;
 
 use icon::Icon;
 use matcher::{Bonus, ClapItem, MatchScope, Matcher};
+use types::{FileNameItem, GrepItem};
 
-pub use self::source::{filter_sequential, SequentialSource};
-pub use self::worker::iterator::dyn_run;
-pub use self::worker::par_iterator::{
+pub use self::parallel_worker::{
     par_dyn_run, par_dyn_run_inprocess, par_dyn_run_list, ParSource, StdioProgressor,
 };
+pub use self::sequential_source::{filter_sequential, SequentialSource};
+pub use self::sequential_worker::dyn_run;
 pub use matcher;
 pub use types::{CaseMatching, MatchedItem, Query, SourceItem};
+
+/// Converts the raw line into a clap item.
+pub(crate) fn try_into_clap_item(matcher: &Matcher, line: String) -> Option<Arc<dyn ClapItem>> {
+    match matcher.match_scope() {
+        MatchScope::GrepLine => {
+            GrepItem::try_new(line).map(|item| Arc::new(item) as Arc<dyn ClapItem>)
+        }
+        MatchScope::FileName => {
+            FileNameItem::try_new(line).map(|item| Arc::new(item) as Arc<dyn ClapItem>)
+        }
+        _ => Some(Arc::new(SourceItem::from(line))),
+    }
+}
 
 /// Context for running the filter.
 #[derive(Debug, Clone, Default)]
